Доработана реализация многопоточных фильтров. Корректная обработка CSRF-токенов. OAuth2 с помощью google id token.
weekend

weekend

weekend

+++ foodsaver/src/main/kotlin/com/foodback/entity/CartEntity.kt
@@ -22,7 +22,7 @@ data class CartEntity(
     var uid: UUID,
 
     @Column(nullable = false)
-    var productCount: Int = 0,
+    var productCount: Long = 0,
 
     @OneToMany(
         fetch = FetchType.LAZY,
